Enable Quote v2 as default
